update the if statement in the generics problem 5

+++ src/problems/generics/problem5_counting_odd_integer_in_collection/CountOddIntegerCollection.java
@@ -10,7 +10,7 @@ public class CountOddIntegerCollection {
         }
         int count = 0;
         for (T element : collection) {
-            if (element % 2 != 0) {
+            if (element.intValue() % 2 != 0) {
                 count++;
             }
         }
